@@ -40,14 +40,14 @@ dependencies:
   path_provider: ^2.1.0
   path: ^1.8.3
   flutter_svg: ^2.0.7
-  amplify_api: ^1.0.0
-  amplify_auth_cognito: ^1.0.0
-  amplify_authenticator: ^1.0.0
-  amplify_storage_s3: ^1.0.0
-  amplify_flutter: ^1.0.0
   go_router: ^6.5.5
   uuid: ^3.0.7
   provider: ^6.0.5
+  flutter_dotenv: ^5.1.0
+  http: ^1.1.0
+  webview_flutter: ^4.2.2
+  shared_preferences: ^2.2.0
+  jwt_decoder: ^2.0.1
 
 dev_dependencies:
   flutter_test:
@@ -73,6 +73,7 @@ flutter:
 
   # To add assets to your application, add an assets section, like this:
   assets:
+    - .env
     - assets/icons/
     - assets/images/
 
