@@ -1,15 +1,14 @@
 import 'dart:io';
 
-import 'package:amplify_auth_cognito/amplify_auth_cognito.dart';
-import 'package:amplify_flutter/amplify_flutter.dart';
 import 'package:camera/camera.dart';
-import 'package:camera_platform_interface/src/types/camera_description.dart';
+import 'package:data_spider/logic/api.dart';
+import 'package:data_spider/logic/login.dart';
 import 'package:data_spider/logic/money_data.dart';
+import 'package:data_spider/models/camera_data.dart';
 import 'package:data_spider/utils/text_styles.dart';
 import 'package:data_spider/utils/theme.dart';
 import 'package:data_spider/widgets/camera_preview.dart';
 import 'package:data_spider/widgets/dialog.dart';
-import 'package:data_spider/widgets/moneyandquantity.dart';
 import 'package:data_spider/widgets/moneytypeandquantity.dart';
 import 'package:data_spider/widgets/sendbutton.dart';
 import 'package:data_spider/widgets/total_amount.dart';
@@ -19,7 +18,6 @@ import 'package:provider/provider.dart';
 import 'package:uuid/uuid.dart';
 
 import 'logic/camera_model.dart';
-import 'logic/provider.dart';
 
 class HomeScreen extends StatefulWidget {
   const HomeScreen({super.key});
@@ -28,23 +26,17 @@ class HomeScreen extends StatefulWidget {
   State<HomeScreen> createState() => _HomeScreenState();
 }
 
-class _HomeScreenState extends State<HomeScreen> {
+class _HomeScreenState extends State<HomeScreen> with WidgetsBindingObserver {
   // providers
-  late HomeScreenProvider _homeProvider;
-  late CameraModel _cameraProvider;
+  // late CameraModel _cameraProvider;
+  // late LoginProvider _loginProvider;
 
   var uuid = const Uuid();
 
   List<MoneyData> moneyWidgets = [];
 
-  String? frontImagePath;
-  String? backImagePath;
-
-  String frontImageTitle = "Нүүр хэсгийн зураг";
-  String backImageTitle = "Арын зураг";
-
-  bool isLoadingFrontImage = false;
-  bool isLoadingBackImage = false;
+  CameraData frontCamera = CameraData(text: 'Нүүр хэсгийн зураг');
+  CameraData backCamera = CameraData(text: 'Арын зураг', isMandatory: false);
 
   TextEditingController totalAmountController = TextEditingController();
 
@@ -57,16 +49,17 @@ class _HomeScreenState extends State<HomeScreen> {
     'imageDesc': [],
   };
 
-  void populateData() async {
-    AuthUser user = context.read<HomeScreenProvider>().user;
-    data['userId'] = user.userId;
-    data['username'] = user.username;
+  bool populateData() {
+    LoginProvider loginProvider = context.read<LoginProvider>();
+    data['userId'] = loginProvider.userId;
+    data['username'] = loginProvider.username;
     data['frontImageId'] = uuid.v4();
     data['backImageId'] = uuid.v4();
     data['totalAmount'] = totalAmountController.text;
     data['imageDesc'] = extractImageDesc();
 
-    validateData();
+    bool isOk = validateData();
+    return isOk ? true : false;
   }
 
   List<Map<String, dynamic>> extractImageDesc() {
@@ -77,27 +70,56 @@ class _HomeScreenState extends State<HomeScreen> {
     return imageDesc;
   }
 
-  void sendData() async {
-    populateData();
-    bool isPublished = await _homeProvider.publish(data);
+  Future<void> sendData() async {
+    print("About to populate data...");
+    bool isOk = populateData();
+    print("Data populated: $isOk");
+
+    if (!isOk) {
+      return;
+    }
+    bool isPublished = await context
+        .read<ApiProvider>()
+        .publish(data, frontCamera.getImagePath, backCamera.getImagePath);
     if (isPublished) {
       alertWidget('Амжилттай илгээлээ', 'Таньд баярлалаа');
       reset();
+      return;
     } else {
-      alertWidget('Амжилттай илгээлээ', 'Таньд баярлалаа');
+      alertWidget('Ямар нэг алдаа гарлаа', 'Та дахин оруулаад илгээнэ үү');
     }
   }
 
   @override
   void initState() {
     super.initState();
-
     moneyWidgets.add(MoneyData());
+  }
+
+  @override
+  void didChangeAppLifecycleState(AppLifecycleState state) {
+    final CameraController cameraController =
+        context.read<CameraModel>().controller;
+
+    // App state changed before we got the chance to initialize.
+    if (!cameraController.value.isInitialized) {
+      return;
+    }
+
+    if (state == AppLifecycleState.inactive) {
+      cameraController.dispose();
+      context.read<CameraModel>().dispose();
+    } else if (state == AppLifecycleState.resumed) {
+      context.read<CameraModel>().initCamera();
+    }
   }
 
   @override
   void dispose() {
     super.dispose();
+
+    totalAmountController.dispose();
+    Provider.of<CameraModel>(context).dispose();
 
     for (var moneyData in moneyWidgets) {
       moneyData.dispose();
@@ -106,31 +128,31 @@ class _HomeScreenState extends State<HomeScreen> {
 
   void reset() {
     setState(() {
-      frontImagePath = null;
-      backImagePath = null;
+      frontCamera.reset();
+      backCamera.reset();
       totalAmountController.text = '';
       moneyWidgets = [];
       moneyWidgets.add(MoneyData());
     });
   }
 
-  void validateData() {
+  bool validateData() {
     if (data['userId'].isEmpty ||
         data['username'].isEmpty ||
         data['frontImageId'].isEmpty ||
         data['totalAmount'].isEmpty ||
-        data['imageDesc'].isEmpty) {
-      // If any of the above fields are empty, show an alert
-      print('========>' + data['totalAmount']);
+        data['imageDesc'].isEmpty ||
+        frontCamera.getImagePath == '') {
       showDialog(
         context: context,
         builder: (BuildContext context) {
           return AlertDialog(
-            title: Text('Дутуу бөглөсөн'),
-            content: Text('* тэмдэглэгээтэй талбаруудыг заавал бөглөнө үү'),
+            title: const Text('Дутуу бөглөсөн'),
+            content:
+                const Text('* тэмдэглэгээтэй талбаруудыг заавал бөглөнө үү'),
             actions: [
               TextButton(
-                child: Text('Ойлголоо'),
+                child: const Text('Ойлголоо'),
                 onPressed: () {
                   Navigator.of(context).pop();
                 },
@@ -139,25 +161,31 @@ class _HomeScreenState extends State<HomeScreen> {
           );
         },
       );
+      return false;
     } else {
+      return true;
       // All required fields are filled
       // Proceed with your logic, for instance, save the data or send it to a server.
     }
   }
 
-  Widget alertWidget(text, content) {
-    return AlertDialog(
-      title: Text(text),
-      content: Text(content),
-      actions: [
-        TextButton(
-          child: Text('Ойлголоо'),
-          onPressed: () {
-            Navigator.of(context).pop();
-          },
-        ),
-      ],
-    );
+  Future alertWidget(text, content) {
+    return showDialog(
+        context: context,
+        builder: (BuildContext context) {
+          return AlertDialog(
+            title: Text(text),
+            content: Text(content),
+            actions: [
+              TextButton(
+                child: const Text('Ойлголоо'),
+                onPressed: () {
+                  Navigator.of(context).pop();
+                },
+              ),
+            ],
+          );
+        });
   }
 
   void _addMoneyWidget() async {
@@ -165,8 +193,6 @@ class _HomeScreenState extends State<HomeScreen> {
     setState(() {
       moneyWidgets.add(moneyData);
     });
-
-    await _homeProvider.postData();
   }
 
   void _removeMoneyWidget() {
@@ -177,42 +203,54 @@ class _HomeScreenState extends State<HomeScreen> {
     }
   }
 
-  void onImageTaken() async {
-    if (frontImagePath == null) {
-      setState(() {
-        isLoadingFrontImage = true;
-      });
-    } else {
-      setState(() {
-        isLoadingBackImage = true;
-      });
-    }
+  Future<void> onImageTaken() async {
+    print("onImageTaken");
+    // if (frontCamera.imagePath == null) {
+    //   setState(() {
+    //     frontCamera.setIsLoading = true;
+    //   });
+    // } else {
+    //   setState(() {
+    //     backCamera.setIsLoading = true;
+    //   });
+    // }
 
-    _cameraProvider.toggle();
-    await _cameraProvider.takePicture((imagePath) {
-      safePrint('imagePath ==================: $imagePath');
-      if (frontImagePath != null) {
+    CameraModel cameraProvider = context.read<CameraModel>();
+    String? imagePath = await cameraProvider.takeImage();
+    if (imagePath != null) {
+      if (frontCamera.imagePath != null) {
         setState(() {
-          backImagePath = imagePath;
-          isLoadingBackImage = false;
+          backCamera.setImagePath = imagePath;
+          backCamera.setIsLoading = false;
         });
-        return;
+      } else {
+        setState(() {
+          frontCamera.setImagePath = imagePath;
+          frontCamera.setIsLoading = false;
+        });
       }
-      setState(() {
-        frontImagePath = imagePath;
-        isLoadingFrontImage = false;
-      });
-    });
+    }
+  }
+
+  void deleteImage(String? path) async {
+    if (path == null) return;
+    File file = File(path);
+    if (await file.exists()) {
+      print("deleting file");
+      await file.delete();
+    }
   }
 
   @override
   Widget build(BuildContext context) {
-    _homeProvider = context.watch<HomeScreenProvider>();
-    _cameraProvider = context.watch<CameraModel>();
+    CameraModel _cameraProvider = context.watch<CameraModel>();
+    // _loginProvider = context.watch<LoginProvider>();
+    Future<void> initializeControllerFuture =
+        _cameraProvider.controller.initialize();
     return Scaffold(
       body: _cameraProvider.isCameraOn
           ? FutureBuilder<void>(
-              future: _cameraProvider.initializeControllerFuture,
+              future: initializeControllerFuture,
               builder: (context, snapshot) {
                 if (!_cameraProvider.isInitialized) {
                   // This means the _initializeControllerFuture hasn't been properly initialized yet
@@ -221,6 +259,7 @@ class _HomeScreenState extends State<HomeScreen> {
 
                 if (snapshot.connectionState == ConnectionState.done) {
                   // If the Future is complete, display the preview.
+                  print("camera on");
                   return Camera(onImageTaken: onImageTaken);
                 } else {
                   // Otherwise, display a loading indicator.
@@ -268,21 +307,22 @@ class _HomeScreenState extends State<HomeScreen> {
                       ),
                       const SizedBox(height: 20),
                       cameraWidget(
-                        frontImageTitle,
+                        frontCamera,
                         onCamera: () {
                           _cameraProvider.toggle();
+                          deleteImage(frontCamera.getImagePath);
                           setState(() {
-                            frontImagePath = null;
+                            frontCamera.setImagePath = null;
                           });
                         },
                       ),
                       cameraWidget(
-                        backImageTitle,
-                        isMandatory: false,
+                        backCamera,
                         onCamera: () {
                           _cameraProvider.toggle();
+                          deleteImage(backCamera.getImagePath);
                           setState(() {
-                            backImagePath = null;
+                            backCamera.setImagePath = null;
                           });
                         },
                       ),
@@ -307,25 +347,15 @@ class _HomeScreenState extends State<HomeScreen> {
     );
   }
 
-  Widget cameraWidget(String text,
-      {bool isMandatory = true, VoidCallback? onCamera}) {
-    String? imagePath;
-    if (text == frontImageTitle) {
-      imagePath = frontImagePath;
-    } else {
-      imagePath = backImagePath;
-    }
-
-    safePrint('imagePath inside cameraWidget ==================: $imagePath');
-
+  Widget cameraWidget(CameraData cameraData, {VoidCallback? onCamera}) {
     return GestureDetector(
       onTap: onCamera,
       child: Column(
         children: [
           Row(
             children: [
-              Text(text, style: TextStyles.textLargeBold()),
-              isMandatory
+              Text(cameraData.getText, style: TextStyles.textLargeBold()),
+              cameraData.isMandatory
                   ? Text("*",
                       style: TextStyles.textLargeBold()
                           .copyWith(color: AppColors.red))
@@ -335,30 +365,29 @@ class _HomeScreenState extends State<HomeScreen> {
           const SizedBox(height: 8),
           Container(
             width: double.infinity,
-            height: imagePath == null ? 200 : 500,
+            height: cameraData.getImagePath == null ? 200 : 500,
             decoration: BoxDecoration(
               color: AppColors.black.withOpacity(0.03),
             ),
-            child: text == frontImageTitle && isLoadingFrontImage
+            child: cameraData.isLoading
                 ? const Center(child: CircularProgressIndicator())
-                : text == backImageTitle && isLoadingBackImage
-                    ? const Center(child: CircularProgressIndicator())
-                    : imagePath == null
-                        ? Center(
-                            child: Column(
-                              mainAxisAlignment: MainAxisAlignment.center,
-                              crossAxisAlignment: CrossAxisAlignment.center,
-                              children: [
-                                SvgPicture.asset('assets/icons/camera.svg'),
-                                Text(
-                                  'Энэд дарж зураг оруулна уу',
-                                  style: TextStyles.textMedium()
-                                      .copyWith(color: AppColors.greyText),
-                                ),
-                              ],
+                : cameraData.getImagePath == null
+                    ? Center(
+                        child: Column(
+                          mainAxisAlignment: MainAxisAlignment.center,
+                          crossAxisAlignment: CrossAxisAlignment.center,
+                          children: [
+                            SvgPicture.asset('assets/icons/camera.svg'),
+                            Text(
+                              'Энэд дарж зураг оруулна уу',
+                              style: TextStyles.textMedium()
+                                  .copyWith(color: AppColors.greyText),
                             ),
-                          )
-                        : Image.file(File(imagePath), fit: BoxFit.contain),
+                          ],
+                        ),
+                      )
+                    : Image.file(File(cameraData.getImagePath!),
+                        fit: BoxFit.contain),
           ),
           const SizedBox(height: 20),
         ],
